pyproject.toml
[tool.poetry]
name = "fintech"
version = "1.0"
description = "a fintech app with django"
authors = ["Namypark"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-unfold = "0.9.0"
asgiref = "3.7.2"
django = "4.2.5"
django-browser-reload = "1.11.0"
sqlparse = "0.4.4"
attrs = "23.1.0"
build = "0.10.0"
cachecontrol = "0.13.1"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
cleo = "2.0.1"
crashtest = "0.4.1"
distlib = "0.3.7"
django-jazzmin = "2.6.0"
dulwich = "0.21.6"
filelock = "3.12.4"
idna = "3.4"
importlib-metadata = "6.7.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jsonschema = "4.17.3"
keyring = "24.2.0"
more-itertools = "10.1.0"
msgpack = "1.0.5"
packaging = "23.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.10.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.5.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
rapidfuzz = "2.15.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.3"
tomlkit = "0.12.1"
trove-classifiers = "2023.8.7"
urllib3 = "2.0.4"
virtualenv = "20.24.5"
xattr = "0.10.1"
zipp = "3.16.2"
django-import-export = "^3.3.1"
django-shortuuidfield = "^0.1.3"
shortuuid = "^1.0.11"
django-money = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
